Start refactoring the persistance module
